Better offset handling for z-fighting
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@52063 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/geogebra3D/euclidian3D/openGL/Renderer.java
@@ -1122,7 +1122,7 @@ public abstract class Renderer {
 	 * @param l
 	 *            the layer
 	 */
-	abstract public void setLayer(float l);
+	abstract public void setLayer(int l);
 
 	// drawing matrix
